perf: avoid eager api client load on startup
